Move shared tests object to their own shared projects. So that AppVeyor doesn't mix up tests.
See #320
