+++ Ninject.Integration.SolrNet.Tests/Ninject.Integration.SolrNet.Tests.csproj
@@ -111,9 +111,9 @@
       <Project>{6E5FBF52-EB19-4F68-AD7E-78DACDDF627D}</Project>
       <Name>Ninject.Integration.SolrNet</Name>
     </ProjectReference>
-    <ProjectReference Include="..\SolrNet.Tests\SolrNet.Tests.csproj">
-      <Project>{F3FE6EF5-CF5C-4461-8691-4A498A463FD5}</Project>
-      <Name>SolrNet.Tests</Name>
+    <ProjectReference Include="..\SolrNet.Tests.Common\SolrNet.Tests.Common.csproj">
+      <Project>{48DA5170-B96E-485C-941D-0D7797DCA0EF}</Project>
+      <Name>SolrNet.Tests.Common</Name>
     </ProjectReference>
     <ProjectReference Include="..\SolrNet\SolrNet.csproj">
       <Project>{CEEB8690-3E08-4440-B647-787A58E71CFA}</Project>
