+++ SolrNet.Tests.Common/Integration.Sample/Product.cs
