+++ SolrNet.Tests.Common/Integration.Sample/ProductLoose.cs
