+++ SolrNet.Tests.Common/MockConnection.cs
