+++ SolrNet.Tests.Common/Mocks/HttpWebRequest.cs
