+++ SolrNet.Tests.Common/Mocks/HttpWebRequestFactory.cs
