+++ SolrNet.Tests.Common/Mocks/HttpWebResponse.cs
