+++ SolrNet.Tests.Common/Mocks/MMappingValidator.cs
