+++ SolrNet.Tests.Common/Mocks/MReadOnlyMappingManager.cs
