+++ SolrNet.Tests.Common/Mocks/MServiceLocator.cs
