+++ SolrNet.Tests.Common/Mocks/MServiceProvider.cs
