+++ SolrNet.Tests.Common/Mocks/MSolrAbstractResponseParser.cs
