+++ SolrNet.Tests.Common/Mocks/MSolrBasicOperations.cs
