+++ SolrNet.Tests.Common/Mocks/MSolrCache.cs
