+++ SolrNet.Tests.Common/Mocks/MSolrConnection.cs
