+++ SolrNet.Tests.Common/Mocks/MSolrDocumentSerializer.cs
