+++ SolrNet.Tests.Common/Mocks/MSolrExtractResponseParser.cs
