+++ SolrNet.Tests.Common/Mocks/MSolrFieldParser.cs
