+++ SolrNet.Tests.Common/Mocks/MSolrHeaderResponseParser.cs
