+++ SolrNet.Tests.Common/Mocks/MSolrMoreLikeThisHandlerQueryResultsParser.cs
