+++ SolrNet.Tests.Common/Mocks/MSolrQuerySerializer.cs
