+++ SolrNet.Tests.Common/Utils/EmbeddedResource.cs
