+++ SolrNet.Tests.Common/Utils/LoggingConnection.cs
