+++ SolrNet.Tests.Common/Utils/Moroco.cs
