+++ SolrNet.Tests.Common/Utils/Node.cs
