+++ SolrNet.Tests.Common/Utils/ProfilerFacility.cs
