+++ SolrNet.Tests.Common/Utils/ProfilingContainer.cs
