+++ SolrNet.Tests.Common/packages.config
@@ -1,0 +1,13 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Castle.Core" version="4.2.1" targetFramework="net46" />
+  <package id="Castle.Windsor" version="4.1.0" targetFramework="net46" />
+  <package id="CommonServiceLocator" version="2.0.1" targetFramework="net46" />
+  <package id="xunit" version="2.3.1" targetFramework="net46" />
+  <package id="xunit.abstractions" version="2.0.1" targetFramework="net46" />
+  <package id="xunit.analyzers" version="0.7.0" targetFramework="net46" />
+  <package id="xunit.assert" version="2.3.1" targetFramework="net46" />
+  <package id="xunit.core" version="2.3.1" targetFramework="net46" />
+  <package id="xunit.extensibility.core" version="2.3.1" targetFramework="net46" />
+  <package id="xunit.extensibility.execution" version="2.3.1" targetFramework="net46" />
+</packages>
